chore: Extrait helper
